Remove manual build system for gunit
All of the build system is now handled by automake. Instead of having
runtests.sh compiling needed Java files for running gunit it's now just
a thing wrapper to parsing the .gunit files. With this change working on
native code goes faster as recompiling the Java files is not needed.