Refer to git log for copyright information
Having to manually keep the year and names updated in each source file
is prone to not remembering to keep it up to date. The same information
can be found from git so just refer people to that. In most places it's
not a requirement to explicitly state such things.