Remove usage of static_pointer_cast
Casting pointers like that is a sign of bast design. We only support
variables at this point so there's no need for abstractions yet.
Variables and functions can't live in the same symbol table any way as
bash allows there to be both a function and a variable with the same
name.