Parser: fix multiple roots in pipeline rule
If you had multiple variable declarations on a single row the pipeline
rule tried to use multiple nodes as the root. This is now fixed via
using the VARIABLE_DEFINITIONS virtual token to group the assignments.
With this change ast_printer -nsf /usr/portage/eclass/*.eclass produces
no error messages. This of course is no guarantee of correct parsing.