Builtin: support unset built-in
We do not support unsetting array indexes currently. In addition,
read-only functions can be unset for now. Note that if no options are
supplied, or the -v option is given, each name refers to a shell
variable(This behavior is supported by bash 3.2, bash-4.1 tries to
unset function if the name doesn't match a variable).