Core: fix function handling
When there handle multiple input files (for example, source built-in is
called while interpreting), there will be multiple token streams. Then
the function index might not be correct as we don't know which stream
the function belongs to. Now both the AST and the function index are
stored to execute the function at the right index in the right stream.
Using single libbashWalker and changing the pointer of the ctns is not
working. The reason might be the internal data structures of
libbashWalker are broken after the function call. So a new libbashWalker
is created every time a function call is made. The current AST is saved
in std::stack because new functions may be defined during a function call.