Core: erase all \newline before parsing
The \newline is treated as a line continuation. Handling \newline in
the hidden channel is not ideal as it might break tokens. For example,
" \newline " will emit two BLANK tokens, which is not what we want.
Now \newline is erased before parsing. The drawback is it might not be
right when handling single quoted string and comments.